pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Maksim Lesnikov <maximborzya75@gmail.com>"]
readme = "README.md"
packages = [{include = "gendifff"}]

[tool.poetry.dependencies]
python = "^3.10"
flake8 = "^6.0.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
coverage = "^7.0.5"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff = "gendifff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
